Scflex-master/Scflex/templates/webserver_template/requirements.txt
aniso8601
beautifulsoup4
blinker
bs4
click
docopt
Flask
Flask-API
Flask-Login
Flask-Mail
flask-mongoengine
Flask-Principal
Flask-RESTful
Flask-WTF
hdfs
itsdangerous
Jinja2
lxml
MarkupSafe
mongoengine
numpy
pandas
passlib
pymongo
python-dateutil
pytz
PyYAML
requests
uWSGI
Werkzeug
WTForms
Flask-Security
htmlmin
